[FEATURE] Implement crawler option "client_config"
This commit adds native support for the recently introduced crawler
options "client_config". Support for this crawler was added in the
upstream package "eliashaeussler/cache-warmup" with version 0.8.2.
Users in need for this crawler option must assure that the latest
version of said upstream package is installed. The new crawler option
is only available to the default crawlers, "ConcurrentUserAgentCrawler"
and "OutputtingUserAgentCrawler". Configuration must be provided as
JSON-encoded string in the appropriate extension configurations
"crawlerOptions" and "verboseCrawlerOptions"