[FEATURE] Allow configuration of recursively accessible pages
User TSconfig for cache warmup may now contain recursively
accessible pages. For this, a plus sign (+) must be added to
a page id to allow cache warmup for all subpages of this page.

Example:

    options.cacheWarmup.allowedPages = 1,72+

In this example, the pages with 1 and 72 as well as all
subpages of page 72 can be warmed up.